+++ src/main/kotlin/net/serverwars/sunsetPlugin/commands/lobby/CommandLobbyLeave.kt
@@ -4,6 +4,7 @@ import com.mojang.brigadier.Command
 import com.mojang.brigadier.context.CommandContext
 import io.papermc.paper.command.brigadier.CommandSourceStack
 import net.serverwars.sunsetPlugin.domain.lobby.exceptions.UpdateLobbyException
+import net.serverwars.sunsetPlugin.domain.lobby.models.Lobby
 import net.serverwars.sunsetPlugin.domain.lobby.models.Participant
 import net.serverwars.sunsetPlugin.domain.lobby.services.LobbyService
 import net.serverwars.sunsetPlugin.translations.sendTranslatedMessage
@@ -24,7 +25,7 @@ object CommandLobbyLeave {
         try {
             val lobby = LobbyService.participantLeaveLobby(Participant(leaver.uniqueId, leaver.name))
             leaver.sendTranslatedMessage("command.lobby.leave.success")
-            lobby.sendMessage("command.lobby.leave.success.notify_lobby", leaver.name, lobby.getParticipantAmount(), lobby.getLobbySettings().size)
+            lobby.sendMessage("command.lobby.leave.success.notify_lobby", leaver.name, lobby.getParticipantAmount(), Lobby.MAX_LOBBY_SIZE)
             return Command.SINGLE_SUCCESS
         } catch (error: UpdateLobbyException) {
             leaver.sendTranslatedMessage(error.key, leaver.name)