+++ src/main/kotlin/net/serverwars/sunsetPlugin/commands/match/CommandMatchEnter.kt
@@ -3,7 +3,10 @@ package net.serverwars.sunsetPlugin.commands.match
 import com.mojang.brigadier.Command
 import com.mojang.brigadier.context.CommandContext
 import io.papermc.paper.command.brigadier.CommandSourceStack
+import net.serverwars.sunsetPlugin.domain.match.services.MatchService
 import net.serverwars.sunsetPlugin.translations.sendTranslatedMessage
+import net.serverwars.sunsetPlugin.util.runAsync
+import net.serverwars.sunsetPlugin.util.runSync
 import net.serverwars.sunsetPlugin.util.sendPlayerToMatch
 import org.bukkit.entity.Player
 
@@ -19,7 +22,13 @@ object CommandMatchEnter {
     }
 
     fun run(joiner: Player): Int {
-        sendPlayerToMatch(joiner)
+        runAsync {
+            if (MatchService.checkInMatch()) {
+                runSync { sendPlayerToMatch(joiner) }
+            } else {
+                joiner.sendTranslatedMessage("command.match.join.error.not_in_match")
+            }
+        }
         return Command.SINGLE_SUCCESS
     }
 }