+++ src/main/kotlin/net/serverwars/sunsetPlugin/domain/lobby/models/LobbySettings.kt
@@ -1,22 +1,16 @@
 package net.serverwars.sunsetPlugin.domain.lobby.models
 
 data class LobbySettings(
-    val size: Int,
     val accessType: LobbyAccessType,
     val gameType: String,
 ) {
     companion object {
-        fun create(size: Int, accessType: LobbyAccessType, gameType: String): LobbySettings {
+        fun create(accessType: LobbyAccessType, gameType: String): LobbySettings {
             return LobbySettings(
-                size = size,
                 accessType = accessType,
                 gameType = gameType,
             )
         }
-    }
-
-    fun withSize(size: Int): LobbySettings {
-        return this.copy(size = size)
     }
 
     fun withAccessType(accessType: LobbyAccessType): LobbySettings {