+++ src/main/kotlin/net/serverwars/sunsetPlugin/listeners/PlayerQuitListener.kt
@@ -1,5 +1,6 @@
 package net.serverwars.sunsetPlugin.listeners
 
+import net.serverwars.sunsetPlugin.domain.lobby.models.Lobby
 import net.serverwars.sunsetPlugin.domain.lobby.models.Participant
 import net.serverwars.sunsetPlugin.domain.lobby.services.LobbyService
 import org.bukkit.event.EventHandler
@@ -12,7 +13,7 @@ object PlayerQuitListener : Listener {
     fun onPlayerQuit(event: PlayerQuitEvent) {
         try {
             val lobby = LobbyService.participantLeaveLobby(Participant(event.player.uniqueId, event.player.name))
-            lobby.sendMessage("command.lobby.leave.success.notify_lobby", event.player.name, lobby.getParticipantAmount(), lobby.getLobbySettings().size)
+            lobby.sendMessage("command.lobby.leave.success.notify_lobby", event.player.name, lobby.getParticipantAmount(), Lobby.MAX_LOBBY_SIZE)
         } catch (_: Exception) {}
     }
 }